refactor : result 패키지를 생성하여 패키지를 분리
