+++ src/main/java/lotto/domain/result/InvalidWinningLottoException.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.result;
 
 public class InvalidWinningLottoException extends IllegalArgumentException{
 	public static final String DUPLICATION = "로또 번호와 보너스 번호는 중복이 될 수 없습니다.";
