+++ src/main/java/lotto/domain/result/LottoRank.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.result;
 
 import lotto.domain.lottomoney.LottoMoney;
